requirements.txt
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
duckduckgo-search>=7.0.0
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
rich>=13.7.0
newspaper3k>=0.2.8
markdownify>=0.14.0
lxml>=5.0.0
lxml_html_clean>=0.4.0
